moved the backend_input message_util into its own folder
