+++ src/thunderbots/software/backend_input/backend.cpp
@@ -1,6 +1,6 @@
 #include "backend.h"
 
-#include "backend_input/message_util.h"
+#include "backend_input/util/ros_messages.h"
 #include "proto/messages_robocup_ssl_detection.pb.h"
 #include "proto/messages_robocup_ssl_geometry.pb.h"
 #include "shared/constants.h"
